Incluyendo algunas validaciones en el controlador y esquema del modelo movie

+++ src/controladores/controladorMovies.js
@@ -17,11 +17,14 @@ async function getMoviesAll (req, res) {
 async function getMoviesById (req, res) {
     try {
         const id = parseInt(req.params.id)
-        const movies = await Movie.findByPk(id, { 
+        const movie = await Movie.findByPk(id, { 
             attributes: ['imagen', 'titulo', 'fecha', 'calificacion'],
             include: Personaje
         })
-        res.json(movies)
+        if (!movie) {
+            return res.json({ error: "No se ha encontrado la pelicula"})
+        }
+        res.json(movie)
     } catch (error) {
         res.json(error)
     }
@@ -29,11 +32,16 @@ async function getMoviesById (req, res) {
 
 async function postMovies (req, res) {
     try {
-        const campos = { ...req.body }
+        const campos = req.body
         if (!(campos.hasOwnProperty('imagen') && campos.hasOwnProperty('titulo') && campos.hasOwnProperty('fecha') && campos.hasOwnProperty('calificacion'))) {
             return res.json({ msg: "Debes pasar la imagen, titulo, fecha y calificacion como minimo para crear un pelicula o serie"})
         }
         const movie = await Movie.create(campos)
+
+        if (!movie) {
+            return res.json({ error: "No se ha podido crear la pelicula"})
+        }
+
         res.json(movie)
     } catch (error) {
         res.json(error)
@@ -43,14 +51,23 @@ async function postMovies (req, res) {
 async function updateMovies (req, res) {
     try {
         const id = parseInt(req.params.id)
-        const campos = { ...req.body }
+
+        const checkExist = await Movie.findByPk(id)
+
+        if (!checkExist) {
+            return res.json({ error: "No se ha encontrado la pelicula"})
+        }
+
+        const campos = req.body
         await Movie.update(campos, {
             where: {
                 id: id
             }
         });
+
         const movie = await Movie.findByPk(id, { 
-            attributes: ['imagen', 'titulo', 'fecha', 'calificacion']
+            attributes: ['imagen', 'titulo', 'fecha', 'calificacion'],
+            include: Personaje
         })
         res.json(movie)
     } catch (error) {
@@ -61,9 +78,18 @@ async function updateMovies (req, res) {
 async function deleteMovies (req, res) {
     try {
         const id = parseInt(req.params.id)
+
+        const checkExist = await Movie.findByPk(id)
+
+        if (!checkExist) {
+            return res.json({ error: "No se ha encontrado la pelicula"})
+        }
+
         const movie = await Movie.findByPk(id, { 
-            attributes: ['imagen', 'titulo', 'fecha', 'calificacion']
+            attributes: ['imagen', 'titulo', 'fecha', 'calificacion'],
+            include: Personaje
         })
+
         await Movie.destroy({
             where: {
                 id: id
